renamed scripts to controllers a little bit more than that and also documented all the routes in the README in the routes folder also removed a test.pdf which was literally 0 bytes
things left todo:
- make db/db.js switching db types from an environment variable
- probably put the current public routes under some form of privacy
- get way more descriptive about
- remove static initialisation files from db and garbage from controllers
- populate test script with actual tests (after a test plan writeup)
