+++ controllers/README.md
@@ -1,0 +1,4 @@
+# Scripts
+
+### About
+The idea of this folder is to have all the scripts, such as parsing scripts and testing scripts and utility scripts and files related to them.
