+++ controllers/eligible_courses_query.js
