+++ controllers/getDegreeProgress.js
