+++ controllers/initialize_quinn.js
