+++ db/README.md
@@ -1,4 +1,4 @@
 # Database Interfaces
 
 ### About
-This includes all the database related scripts. The amount of files here will decrease
+This includes all the database related scripts. There are also some sql files, such as `database_query.sql` and `wip_accurate_eligible_courses.sql` which really serves as documentation to how various requests are run from an SQL point of view
