+++ db/wip_accurate_eligible_courses.sql
